refactor: extract StatCards from useStatCards to component
