+++ client/src/modules/CourseStatistics/components/StatCards/StatCards.tsx
@@ -1,19 +1,30 @@
 import { CoursesStatsDto } from '@client/api';
-import { StudentsCountriesCard } from '../components/StudentsCountriesCard';
-import { StudentsStatsCard } from '../components/StudentsStatsCard';
-import { MentorsCountriesCard } from '../components/MentorsCountriesCard/MentorsCountriesCard';
-import { EpamMentorsStatsCard } from '../components/EpamMentorsStatsCard';
-import { StudentsWithMentorsCard } from '../components/StudentsWithMentorsCard';
-import { StudentsWithCertificateCard } from '../components/StudentsWithCertificateCard';
-import { StudentsEligibleForCertificationCard } from '../components/StudentsEligibleForCertificationCard';
-import { TaskPerformanceCard } from '../components/TaskPerformanceCard';
-import { StudentsCertificatesCountriesCard } from '../components/StudentsCertificatesCountriesCard';
+import { StudentsCountriesCard } from '@client/modules/CourseStatistics/components/StudentsCountriesCard';
+import { StudentsStatsCard } from '@client/modules/CourseStatistics/components/StudentsStatsCard';
+import { MentorsCountriesCard } from '@client/modules/CourseStatistics/components/MentorsCountriesCard/MentorsCountriesCard';
+import { EpamMentorsStatsCard } from '@client/modules/CourseStatistics/components/EpamMentorsStatsCard';
+import { StudentsWithMentorsCard } from '@client/modules/CourseStatistics/components/StudentsWithMentorsCard';
+import { StudentsWithCertificateCard } from '@client/modules/CourseStatistics/components/StudentsWithCertificateCard';
+import { StudentsEligibleForCertificationCard } from '@client/modules/CourseStatistics/components/StudentsEligibleForCertificationCard';
+import { TaskPerformanceCard } from '@client/modules/CourseStatistics/components/TaskPerformanceCard';
+import { StudentsCertificatesCountriesCard } from '@client/modules/CourseStatistics/components/StudentsCertificatesCountriesCard';
+import Masonry from 'react-masonry-css';
+import css from 'styled-jsx/css';
 
-type UseStatCardsProps = {
+type StatCardsProps = {
   coursesData?: CoursesStatsDto;
 };
 
-export function useStatCards({ coursesData }: UseStatCardsProps) {
+const gapSize = 24;
+
+const masonryBreakPoints = {
+  default: 4,
+  1100: 3,
+  700: 2,
+  500: 1,
+};
+
+export function StatCards({ coursesData }: StatCardsProps) {
   const cards = [
     coursesData?.studentsCountries && {
       title: 'studentsCountriesCard',
@@ -71,5 +82,36 @@ export function useStatCards({ coursesData }: UseStatCardsProps) {
       },
   ].filter(Boolean);
 
-  return { cards };
+  return (
+    <>
+      <Masonry
+        breakpointCols={masonryBreakPoints}
+        className={masonryClassName}
+        columnClassName={masonryColumnClassName}
+      >
+        {cards.map(({ title, component }) => (
+          <div style={{ marginBottom: gapSize }} key={title}>
+            {component}
+          </div>
+        ))}
+      </Masonry>
+      {masonryStyles}
+      {masonryColumnStyles}
+    </>
+  );
 }
+
+const { className: masonryClassName, styles: masonryStyles } = css.resolve`
+  div {
+    display: flex;
+    margin-left: -${gapSize}px;
+    width: auto;
+    min-height: 85vh;
+  }
+`;
+const { className: masonryColumnClassName, styles: masonryColumnStyles } = css.resolve`
+  div {
+    padding-left: ${gapSize}px;
+    background-clip: padding-box;
+  }
+`;
