+++ client/src/modules/CourseStatistics/components/StatCards/index.ts
@@ -1,0 +1,1 @@
+export { StatCards } from './StatCards';
