+++ client/src/modules/CourseStatistics/hooks/index.ts
@@ -1,2 +1,1 @@
 export { useCoursesStats } from './useCourseStats';
-export { useStatCards } from './useStatCards';
